+++ src/steps/scans/index.ts
@@ -14,7 +14,7 @@ import {
   ScanStatus,
   User,
   VulnerabilityExport,
-} from '../../tenable/types';
+} from '@jupiterone/tenable-client-nodejs';
 import {
   createScanEntity,
   createScanFindingRelationship,